refactor: rename and reorganize user action test classes
